pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relaxed-recursive-transformers"
version = "0.1.0"
description = "JAX/Flax implementation of Relaxed Recursive Transformers (ICLR 2025)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ahmed Taha", email = "at3889@columbia.edu"},
]
keywords = ["jax", "flax", "transformers", "recursive", "lora", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "flax>=0.8.0",
    "optax>=0.1.7",
    "orbax-checkpoint>=0.5.0",
    "transformers>=4.38.0",
    "datasets>=2.16.0",
    "tqdm",
    "pyyaml",
    "einops",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
